+++ commands/reputation/rep.js
@@ -1,5 +1,5 @@
 const discord = require('discord.js');
-module.exports.run = (msg, Flags, Parameters, bot, ret)=>{
+module.exports.run = async (msg, Flags, Parameters, bot, ret)=>{
     let member = Parameters.get('member');
     if(!Parameters.isSet('member')) {
         member = msg.member;