+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # CHANGELOG
+## 01/03/2021 - Monday - Version 0.8.2 - Reputation patch
+- Members can no more give reputation to self.
+- Cooldown actually works
+- Added reptop, the reputation top 3 for server(number can be changed using parameter -c)
 ## 01/03/2021 - Monday - Version 0.8.1 - Calendar Patch
 - Fixed calendar displaying wrong month
 - Calendar now displays dates in order